Added Main file for the project
Project is very simple in scope uses python OS library to interface with the API to detect batter status information and details regarding various statistics no longer displayed in the Windows 10/11 battery context menu. 

There are a few finer details to change up such as the creation of a log directory.