Altered main.py
In the initial startup phase, main.py now checks for the existence of a local logs directory in same location as main.py. (Later this could be added to a configuration file or element in the code, but that may bypass the root of this being a simple scale project).